make `goto` exceptions the default
Summary
=======

This is a preparation for the removal of the `setjmp` and `quirky`
exceptions modes. `goto` exceptions were already the default when using
`--gc:arc|orc`.

* fix the programs crashing when a thread was created and both thread-
  local storage emulation and `--exceptions:goto` were enabled
* fix execution continuing after a call to an RVO-using closure when
  using `--exceptions:goto`
* fix an issue where assignments were observable if the right-hand
  expression is a call that raised an exception
* remove the `segfaults` standard library module; its current
  implementation is incompatible with anything besides setjmp-exceptions
* `--exceptions:native` now means `--exceptions:goto` when using the C
  target

There's currently an issue with the ordering of routines in the `system`
module, which prevents `nimrtl` from compiling when using
`--exceptions:goto`.

Details
=======

The infinite recursion error that occurred when using emulated thread-
local storage is caused by the procedure for accessing emulated
threadvars (`GetThreadLocalVars`) getting an error flag query (read,
access of a threadvar) injected, now that `system` procedures are no
longer treated as never raising a defect.

The intermediate solution: disable error flag injection for all
`.compilerproc`s. As a consequence, exceptional control-flow inside
their  immediate bodies is no longer possible. In order to still support
raise-dispatcher-like compilerprocs (e.g. `raiseFloatOverflow`), an
unconditional return is inserted after the call to a `noreturn`
procedure when inside a procedure that has the error flag disabled.

In addition, special-case the `threadProcWrapper` system procedure to
also have the the error flag (and thus exceptional control-flow)
disabled, as it is responsible for setting up the emulated thread-local
storage.

Standard Library
----------------

The `segfaults` module implicitly relied on `setjmp`-exceptions, as
only `longjmp` is able to leave the signal handler and enter the
NimSkull exception handler without additional support from the code
generator.

Since it would become non-functional with the planned removal of
`setjmp`-exceptions, it is removed already.