remove `setjmp` and `quirky` exception modes
Summary
=======

They're removed in order to shrink the amount of language dialects,
and to slightly simplify the compiler. The only remaining exception
handling mode for the C target is now `goto`.

* remove the `nimQuirky` define
* remove the `setjmp` and `quirky` exceptions modes

In addition, remove the C bindings for `longjmp`, `sigsetjmp`, and
`__builtin_setjmp` from the `ansi_c` module, plus the related built-in
defines. The `setjmp` procedure is still used by the GCs employing
stack-scanning, and thus has to be kept for now.

Details
=======

* build koch and the first bootstrap iteration with
  `--exceptions:goto`, so that they continue to compile
* remove `setjmp` and `quirky` exceptions related code from the code
  generator and semantic analysis

Runtime Library
---------------

Remove `nimLeaveFinally`, `pushSafePoint`, `popSafePoint`, and
everything related to `TSafePoint`, including the type itself.