cgen: handle uncaught exceptions at module exit
Summary
=======

For the C target, uncaught exceptions were only detected and reported
(via `nimTestErrorFlag`) when raised from inside the main module, but
not when raised from inside others.

When starting to execute the next module, the error mode was still
active, which, depending on the contents of the module's module-level
code, could lead to a cascade of failures eventually resulting in a
crash unrelated to the original exception.

A call to `nimTestErrorFlag` is now inserted at the exit of every
module's init procedure.

Misc
====

- fix a `FieldDefect` error in `vmrunner` that ocurred when reporting an
  unhandled exception