cfg: drop project ".nimcfg" support
It was possible to defined a project config with a `.nimcfg` extension,
this is no longer supported. Now config file discovery is a bit simpler
and this half-baked bug-in-waiting is removed.

`nimconf` now uses a simple iterator to produce a list of config files
to attempt processing, making its implementation easier to follow.
Removed ".nimcfg" from `options.findProjectNimFile`'s heuristic.