drop `ExternalReport` dep in `extccomp` module
This removes `ExternalReport`, legacy report cruft, from the `extccomp`
module.

Made `extccomp.setCC` proc no longer report an error, instead return the
compiler that it set. Callers now deal with the result as required, in
this case `commands` and `main` both need to emit errors. These errors
have to temporarily use legacy report but that's temporary. As part of
this change `listCCnames` was exported in order to allow querying of
compiler definitions, which seems like an oversight all along.