compiler: compile stats on success refactor (#659)
## Summary

Compiler states generated on successful compile are now generated in
`main`. This removes the dependencies on legacy reports and msgs.

## Details

This is mostly a refactoring undoing poor design after worse, the
message is now generated and emitted via `main`, removing this code from
`msgs` and "reports" modules. In addition, to the above fixed a bug
where VM profiling dump was not produced for unsuccessful compiles.