astgen: fix issues with the MIR -> AST step
Summary
=======

Fix two issues in `astgen` where the semantics of the MIR were not
correctly mapped to `PNode` AST. This led to `wasMoved` operations only
being applied to intermediate temporary locations in some situations,
causing double frees.

Details
=======

The first issue was that an expression like `(a; lvalue).c` was not
detected to be an lvalue expression due to the first operand being a
statement list expression (`nkStmtListExpr`), causing a temporary
instead of an lvalue reference (view) to be created when using the
expression as a to-be-consumed operand to a MIR region.

If the operand was passed on to a `wasMoved` operation, only the
temporary location is reset, but not the expected source location. To
visualize the problem:
```nim
  x = callReturningASeq()[0] # x is something with a destructor
  # gets transformed to MIR code that, after translation back to AST,
  # looked roughly like:
  var tmp: T
  var tmp2: T = (tmp = callReturningASeq(); tmp)[0]
  x = tmp2
  wasMoved(tmp2) # `tmp[0]` is what needs to be reset here
  ...
```

For fixing the issue, the existing unwrapping transformation that is
applied prior to detecting whether an expression is a view is extended
to also "forward" nodes: an expression like `(a; b).c` gets transformed
into `(a; b.c)`. Besides fixing the issue, this also reduces nesting of
the generated AST a bit.

Using the example from above, it now looks like:
```nim
  var tmp: T
  tmp = callReturningASeq()
  var tmp2: var T = tmp[0]
  x = tmp2
  wasMoved(tmp2)
  ...
```
which is correct.

The second issue was with 'consume' arguments of numeric type not being
passed to a region by name (if possible).