improve `getAst` magic processing (#664)
## Summary

Move part of the processing into `transf` and transform the magic call
into something that matches its semantics better. This fixes a small
issues where the `getAst` magic appeared to return nothing when the
meta-routine to expand is a macro.

## Details

A `getAst` expression like `getAst(templ(a, b, c))` is now transformed
into `getAst(templ, a, b, c)` during `transf`. This is slightly more
correct at the MIR level, as `templ` is not "called" here, but is rather
an argument to the `getAst` magic.

* remove the `mirgen` special handling of `static` parameters for
  templates and macros in a `getAst` context. Attempting to work around
  the issues with `getAst` like this is not a good idea
* adjust `vmgen` to work to the new flat representation of `getAst`
  calls
* remove the `vmGenDiagBadExpandToAstArgRequired` diagnostic (semantic
  analysis already rejects ill-formed `getAst` argument expressions)

The regression was due to `vmgen.genCall` using the type of the
macro call expression (which incorrectly was `nil` when the macro had no
return type specified) for detecting whether it returns something.