vmgen: work around a VM issue (#649)
## Summary

Adjust the VM code generator to always use an intermediate temporary
when assigning the result of a call to a local of primitive type. This
fixes wrong values or VM crashes when raising from inside the procedure
appearing on the right-hand side of such assignments.

The test suite was affected (`lang_stmts/casestmt/tcasestmt.nim`), so
this is another step of making the test suite work with a `release`
compiler.

## Details

The call-and-assign `IndCallAsgn` operation is currently implemented by
moving the register representing the destination register into the
callee's register file (its stack frame). Once the procedure exits, the
register is moved back, but only if the procedure didn't exit via an
unhandled exception.

While this is not a problem for complex return values (a temporary is
currently always used for them), it is for primitive ones (int, float,
etc.) stored directly in registers, as no temporary is used for them and
thus raising an exception from inside the callee leaves the register
empty.

There are other workarounds possible, but adjusting the code generator
is the one with the smallest amount of impact.