cli: now properly exits if no mainmodule exists (#655)
## Summary

This fixes a legacy reports bug, where the migration stopped program
termination even when requiring a main module to exist. This now works
and a more approriate error message is provided instead of a path error.

## Details

`modules.wantMainModule` now generates an `rextCmdRequiresFile` (new
legacy report kind), which is treated as a fatal error by
`options.isCompilerFatal`. In addition a quick tests was added to ensure
this continues to function going forward.