fix string values for legacy reports enums
Summary

The string values are what end up in the `[Foo]` suffix of error
messages; and used to refer to them in options `--hint[foo]:off`. These
were missed when initially adding the now legacy reports system.

Details

The changes were as follows:
- `semHintLibDependency = "Dependency"`
- `rcmdRunnableExamplesSuccess = "Success"`

Minor comment update for the `ExternalReport` category.