cli: msgformat is for debug/test only
Summary

The `--msgformat` legacy reports 'feature' causes too many issues,
downgrading it to a dev/debug/test only "feature". This is now only
available for the `nim` compiler executable; it should be restricted
further.

Details

It can now only be set as a cli flag, not via config; docs reflect this.
In addition, since legacy reports conflates, formatting, recovery, and
everything else, this format setting was moved to `cmdlinehelper`, which
uses a callback on `ConfigRef` to update the structured report hook...
for formatting.

This dropped `commands` dependency on cli/sexp reporters, and introduced
an enum, `MsgFormatKind` to encapsulate the choices. Great care hasn't
been taken with this as it's legacy reports cruft.