nimblecmd: remove legacy reports (#670)
## Summary

Although nimble itself is likely to be removed, simplifying the code to
enable throwing out reports from CLI/command modules.

## Details

`nimblecmd` module now returns a data type of what it did, then 
`commands` processes this and determins what to output if at all.

In addition, the `tnimblecmd` test had to be updated as the API no
longer accepts `TLineInfo`, which was only required for "reports".
Additionally, this fixed a minor bug where messages generated lacked
line info, a regression introduced with legacy reports.