mirgen: ignore `.compileTime` variables
Summary
=======

Don't emit a 'def' for `.compileTime` globals in `mirgen`. Until the
implementation of `.compileTime` variables gets an overhaul, this
restores their lazy-generation behaviour when destructors are used, but
also means that they're for now not destroyed anymore when going out of
scope.

Details
=======

Lazy-generation not working when destructors are used led to issues when
the `.compileTime` variable in question has an initializer expression
that makes use of compile-time-only procedures.

* ignore definitions of `.compileTime` variables outside of non-compile-
  time contexts in `mirgen`
* remove now obsolete `.compileTime` variable handling from `cgen` and
  `jsgen`
* run the `tunique_type.nim` test with `--gc:orc`

Registering destructors for `.compileTime` variables that require them
is possible, but skipped here because it would only introduce more
obsolete-on-arrival code.