tests: fix the `tpackedsets.nim` test
The `typed` template parameter was used in a definition context (likely an
oversight). This lead to the local being treated as a procedure-level
global that required destruction.

A global destructor was thus registered, but since the global is not a
top-level global and destructor call for it was already injected, this
resulted in a double free at run-time.