internal refactor of hint/warn flag processing (#673)
## Summary

Hints and warnings can be enabled/disabled via `--hint[foo]:on` on the
CLI or in cfg files, this change separates the logic from legacy reports
in the internal implementation.

## Details

The `processSpecificNote` procedure within the `commands` module handles
this processing, introduced a revised version which returns a result
data type describing the outcome of processing. Then used a template to
bridge the result to legacy reports output. The core option processing
in `processSwitch` then uses the template making this change
transparent. Also fixed a minor issue where line info wasn't being
passed in when cfg files were involved.