renabled `js` testing (#676)
## Summary

`js` category tests were disabled due to a bug in testament. This change
enables them. As part of this `jsffi.bindMethod` was found to broken and
was deleted, the author fundamentally misunderstands JavaScript's `this`
binding and `bindMethod` just doesn't make sense.

## Details

The bug in testament was boolean logic which disabled the entire `js`
category in the `processCategory` proc. After fixing this a few updates
were required:
- deleted `bindMethod` the idea is fundamentally wrong
- fix tests using deleted `table` API `add`
- fix malformed test spec in `tjs_bitwise_operator_64bit`