system: more correct `SysThread` bindings
Summary
=======

Change the `pthread_t` type binding (`SysThread`) to be treated as a
`distinct pointer` instead of an `object` for all targets besides 64-bit
Linux at the NimSkull side. The Posix standard doesn't require
`pthread_t` to be implemented as a pointer, but many implementations do
implement it as an opaque pointer.

Details
=======

Treating the type as an `object` led to the C code generator using
bracket initialization for initializing it in certain contexts, which is
incorrect if `pthread_t` is not implemented as a C struct and thus lead
to C compiler errors (MacOS was affected when using `--gc:orc`).