consider destructors of concrete types bound to concepts
Summary
=======

Turn the `hasDestructor` template into a procedure and skip wrapper-like
types (including `tyUserTypeClasses`) before testing for the
`tfHasAsgn`. This fixes:

* resolved concept types not being considered by cursor inference,
  destructor injection, or the move analyser
* `supportsCopyMem` reporting 'false' for concept types where the bound
  concrete type has a destructor

In addition, the JS backend failing with an internal compiler error for
variables using a resolve type-class as the type is also fixed.

Details
=======

Since the `tyUserTypeClasses` set is not included in `abstractInst`, the
logic part of the mid- and back-end needs to skip them manually, which
it more often than not does not. A `knownIssue` test is added for one
case where this leads to an issue.