internal: clarify type-bound operator attachment (#680)
## Summary

Move to only attaching type-bound operators to unwrapped types, by
skipping to the underlying concrete type before calling
`createTypeBoundOps` in `sempass2`. This is small step towards cleaning
up the type-bound operator implementation.

In addition, this also fixes a regression caused by `hasDestructor` now
skipping the wrapper-like types, which caused alias types and generic
instantiations to not be detected as having a destructor attached.

## Details

* move the skip set used in `hasDestructor` to the `skipForHooks`
  constant
* skip types from the `skipForHooks` set before querying or creating the
  type-bound operators for a type in `sempass2`
* skip types from the `skipForHooks` set before querying type-bound
  operators in `injectdestructors`
* don't attempt to rewrite explicit type-bound operators call during
  `sem` -- that's the responsibility of the `sempass2`
* use `getOp` for querying the attached sink operator in
  `injectdestructors` (it should have been used there from the start)

Calls to `createTypeBoundOps` from `liftdestructors` still attach the
operators to the original (read, unskipped) type.

### The fixed problem

Both closure types (`tyProc` with the `ccClosure` calling convention)
and compound types where none of the element/field types has an
overridden type-bound operation at the time when the compound type is
defined don't have the `tfHasAsgn` flag included at definition time.

`createTypeBoundOps` only includes the flag on the type that was passed
to it, not the skipped one, meaning that if `createTypeBoundOps` is
never invoked directly with the skipped type, the skipped type doesn't
get the `tfHasAsgn` flag included, resulting in it being effectively
treated as having no attached type operation.