compiler: default to `--gc:orc` for the C target (#630)
## Summary

After bootstrapping the compiler itself with `--gc:orc`, this is the
second step towards removing the old runtime.

In order to make to make sure that ORC is stable enough, and that the
standard library works with the new runtime, `--gc:orc` is now implied
for the C target when no memory management strategy is explicitly
selected.

### The New Runtime

Using either `--gc:orc` or `--gc:arc` implies usage of the new runtime,
meaning that the new runtime is now the default. The key differences
compared to the old runtime are:
* usage of `deepcopy` in alive code requires the `--deepcopy` option to
  be enabled
* modifying a string through unsafe facilities (pointers and `cstring`)
  requires preparing the string for mutation first (via
  `prepareMutation`)
* `GC_unref` and `GC_ref` are no longer available for `string` and `seq`
  types
* `repr` version 2 is used

With `--gc:orc` and `--gc:arc`, `shallowCopy` is a no-op.

## Details

* support `prepareMutation` for code running in the VM (it's currently a
  no-op)
* fix a crash with `marshal.loadAny` and empty strings, when using the
  new runtime
* disable tests depending on `nimrtl` -- `nimrtl` doesn't yet work with
  the new runtime
* the IC implementation doesn't support type-bound operators (and thus
  `--gc:arc|orc`) at this time, so change its tests to explicitly use
  `--gc:refc`
* adjust the test suite where necessary