IC: fix the code generator crashing
Summary
=======

Always allocate space for all modules in the backend module list before
starting code generation, regardless of what parts of the list are
filled. This fixes the C code generator crashing with an `IndexDefect`
in certain situation where not all modules were recompiled.

Details
=======

If a module is cached there's no change in its set of alive symbols, no
backend module list entry is created for it. No entry being present is
not a problem for the code generator, but the module list still needs to
have an empty slot reserved for the entry.