cgen: fix missing definitions
When using alive data from the new DCE implementation, only non-magic
routines or routines part of a select set of magics are considered for
code generation. The magics used by the seqsv2 implementation were not
included there, so no code was generated for their instantiations.

As a preparation for the planned backend processing unification, apart
from being adjusted, the set is also split out into a constant.