sem: process `mEnumToStr` earlier
Summary
=======

Mark compiler-generated enum-to-string procedures as being `mEnumToStr`
instances, and replace the generic `mEnumToStr` magic symbols with the
compiler-generated ones during semantic analysis already.

This makes it easier for code generators to opt-in to using the
compiler-generated procedure instead of implementing it via custom
logic. As an intentional side-effect of this change, the default
`$` operator for enums can now be used when both the IC backend and
new runtime are enabled.