static `try` void-expression no longer crashes (#682)
## Summary

Fix compiler crash for `try` void expressions in `static` blocks;
moreover, this fixes a crash for any AST that produced bytecode but
didn't allocate any registers.

## Details

Static blocks presently are void-expressions (statements), `vmgen` and
the `compilerbridge` were not handling these correctly in all cases.
When `vmgen` generated bytecode that didn't allocate any registers the
`compilerbridge` would crash attempting to access a non-existent local
for a return value. One way to trigger this scenario is with some void-
expressions (see added test).

In particular it uses `opcRet` for expression evaluation, where a return
value is required from VM code execution, and `opcEof` when one is not
required. In the case of the former the present behaviour is largely
maintained except we now return the index of local storage in the main
VM stack-frame; and for the latter we don't return any index signaling
to the consumer that no value as a `TFullRegister` with kind `rkNone` is
to be provided.

In debugging this change, an issue with VM code gen listings was found
and resolved, as well.