tester: escape strings in EPC mode
Summary
=======

Escape the path strings part of the S-expression messages generated by
the `nimsuggest` tester. Because of the backslashes used by paths on
Windows, the strings not being escaped properly led to all EPC tests
failing there.

Details
=======

Since the strings are part of an S-expression, `escapeJson` is used
instead of `escape`.