front: centralize backend-specific customization
Summary
=======

Start moving backend-specific customization into a single procedure that
is called prior to executing a command (now including `nimsuggest`). In
addition, fix `nim e` configuring the compiler as if the C backend was
used -- defines like `c` are now no longer present when running
NimScript.

This is another step towards cleaning up the front-end, and fixes the
following issues:
- the backend defines (`c`, `js`, `vm`) not being present in
  `nimsuggest` mode
- no GC being selected by default in `nimsuggest` mode
- the incorrect exception mode being selected when using `nimsuggest` or
  `check`

Details
=======

* move `customizeForBackend` into a standalone procedure in currently
  located in `main` (until a better home for it is found) and call it
  before running `nimsuggest`'s `mainCommand`
* move the exception mode configuration for the JS and VM backend into
  `customizeForBackend`
* don't call `customizeForBackend` prior to executing a file in
  NimScript mode
* fix the GC staying unselected when using the JS or VM backend
* remove the now obsolete `selectDefaultGC` procedure

Due a side-effect of using the new runtime, the `tsigmatch.nim` test
fails failed because of an issue with type aliases in procedure
instantiations. A `knownIssue` test is added.