reports: remove most CLI/cfg related legacy (#689)
## Summary

With the introduction of `commands.CliEvent` and compiler messages a
number of legacy reports related to thses are no longer required. This
change removes a large majority of these.

## Details

Removes a number of report kind enums, most of these are external
reports relating to command, flags, and arguments. This also removes a
few incorrectly categorized lexer reports, which relate to cfg
processing.

This change is entirely internal, and the messages are still produced as
before.