IC: support type-bound operators and `--gc:orc` (#685)
## Summary

Finish the implementation for recording type-bound operator into packed
modules, and adjust the IC backend and its DCE implementation to support
both lifetime hooks and parts of the new runtime.

This allows for simple programs to be compiled with `--incremental:on`
and `--gc:orc`, and is another step towards removing the old runtime.

## Details

Getting type-bound operators and `--gc:orc` supported consists of four
parts.

### Packed Modules

Attaching either a user-defined or generated type-bound operator to a
type is a change to the `ModuleGraph`, and thus has to be recorded in
order to allow for later replay of the state changes. The basic
facilities already existed, but storing and replaying the `attachedOps`
changes was missing.

* implement recording the attachment changes
* implement replaying the attachment changes
* split out replay of state relevant to the backend into a separate
  procedure
* fix the "trivial attached operator" check not taking on-demand
  loading of procedure bodies into account

### Backend setup

Restore backend-relevant `ModuleGraph` state discarded by
`resetForBackend` before running alive analysis -- no information
about attached operator would be present otherwise.

### Dead-code elimination

Compared to the other backends, the IC implementation uses a dedicated
alive analysis pass over the whole packed module-graph in order to
figure out the set of alive routines. Only routines present in this set
have C code generated for them -- the others are ignored.

The `injectdestructors` pass, which is responsible for injecting copy,
sink, and destroy hooks, is only run for all alive code reaching the
code generators, but at that point, the previous alive analysis would
have marked all not explicitly used hooks as dead code.

In order to get around this phase-ordering problem, the alive analysis
now conservatively marks all lifetime hooks for types used by alive:
* locals
* globals
* constants
* parameters

as being in the alive set. The same goes for lifetime hooks attached to
types used as the return type of alive routines.

The downside of this conservative approach is, that hooks ending up
not being actually used are still emitted and compiled, leading to
larger binaries.

### Code generator

Include the magics used by the seqsv2 implementation in the set of
magics to consider for code generation. Said magics are implemented
as normal (generic) procedures, and thus their bodies also need to
be compiled.