compiler: add the `native` GC option (#692)
## Summary

Introduce the `native` GC option and make it the default for the
JavaScript and VM targets. `refc` was previously selected there, but
this is not correct, as the JavaScript target uses JavaScript's garbage
collector and the VM uses its own internal one.

Besides allowing for distinguishing between `refc` and "native GC", this
is also a preparation for the removal of `refc`, as without the `native`
option, both targets would be left without a GC option to use internally
until they support ORC.

For the C backend, `--gc:native` redirects to `--gc:orc`.