Rename the `native` gc option to `default` Summary ======= Rename the new `native` gc option to `default` for improved clarity. Made the `native` option internal only.
Details
=======
I raised a discussion about the naming of the new `native` gc option
due to it potentially being confusing depending on the backend.
We agreed to rename it to `default` to better fit what this option
actually means. I made the `native` option not choosable via
command-line switches as it should be an internal implementation detail.