system: declare `prepareMutation` for all targets
Summary
=======

The system procedure for preparing a string for mutation
(`prepareMutation`) was previously only available when targeting C,
requiring one to guard usages of it with `when` statements.

Make it available for all targets, with the ones not using copy-on-write
string literals treating it as a no-op. This makes it easier to write
code that works without change across all backend.

Details
=======

* declare `prepareMutation` for the JavaScript and VM target
* remove `when` guards around `prepareMutation` call in the standard
  library and tests