Make the new `native` gc internal only (#696)
## Summary
Made the new `native` gc not accessible via command line arguments or
via `compileOption`.

## Details
The latest `native` gc was created to better represent the gc the JS and
VM backends use. But it has no meaning for the C backend and instead
redirected to orc. As there is no reason to explicitly choose the
`native` gc via command-line arguments I removed the option. I also
removed the option for the `compileOption` proc because one should
query the used backend instead.