cgen: remove `optSeqDestructors` usage
The option indicates whether version 2 sequences are enabled, which is
always the case for the C backend after the legacy GCs are removed.

Note that the flag was misused for checking whether a destructor-based
GC is enabled, meaning that some of the removals here are more related
to the GC removal than to the seq v1 removal.