koch: fix the ORC bootstrapping test
The first iteration unconditionally used `--gc:refc`, which is a problem
when the bootstrapping compiler is one that no longer supports
`--gc:refc`.

The newer compiler versions (which properly support `--gc:orc`,
although not necessarily default to it) identify themselves as "Nimskull
Compiler" (compared to "Nim Compiler"), and `koch` now uses this
information in order to decide whether `--gc:refc` needs to be
enabled.