cli: remove the `--seqsv2` option
It's purpose is to only enable the destructor-based `seq`s and
`string`s, but nothing else. The destructor-based one is now the default
(and only) available implementation for the C backend, and since both
the JavaScript and VM backend don't support destructor-based sequences
yet and would fail to compile if forced to use it, the option is
removed.