remove legacy condsyms
This change removes a number of the old defines that were used for
compiler and/or standard library change detection.

There were a number of haphazardly named defines for a half-baked system
to detect if certain changes were made to the compiler or standard
library. The approach wasn't good, the design leads to lots of action at
a distance, and just more stuff to otherwise puzzle over.