compiler: remove support for the legacy GCs (#549)
## Summary

There are multiple issues with the legacy memory management strategies
(the ones that are not destructor-based):

* each one introduces another language dialect
* they necessitate the old runtime (more code to maintain)
* they're specific to the C backend, making them not fit in well with
  the direction of having each backend support the full language
* the way they're currently integrated with the compiler introduces a
  large amount of complexity and conditional logic to the C code-
  generator

For the reasons listed above, compiler support for both the legacy GCs
and large parts of the old runtime is removed. The only remaining
selectable GC modes are `orc` and `arc`.

Except for the v1 run-time-type-information, which is still required by
the `deepcopy` implementation and `marshal` module, the old runtime is
now obsolete and can be removed.

## Details

**CLI:**

* adjust the allowed `gc` switch arguments to only include
  `destructors`, `arc`, and `orc`
* remove the obsolete `--seqsv2` option (the version 2 implementation is
  always used now, at least for the C backend)

**C code-generator:**

* remove all usages of both `optSeqDestructors` and `optTinyRtti` --
  they're now unconditionally enabled when using the C backend
* remove `optTinyRtti`, as it has no remaining usage-sites
* remove all code related to the old runtime's `string`s and `seq`s
* remove all support for the garbage collectors using the old runtime
* remove code-generation logic for the unused magics `mIntToStr`,
  `mInt64ToStr`, and `mFloatToStr`
* simplify the code generated for the entry point

**Tests:**

* remove tests specific to the removed GCs
* disable the tests for `--gc:none`
* don't specify `--gc:arc` for some tests where not required, allowing
  the tests to be joined

**Documentation:**

* clean up the memory management documentation (`gc.rst`)
* remove some mentions of the removed GCs and the old runtime

**CI:**

* remove `boehm`-related package dependencies

**Misc:**

* remove `when` statements related to the removed GCs from compiler code
* support `koch` building the first iteration with `--gc:orc`, which is
  a preparation for updating the csources compiler to one that doesn't
  support `--gc:refc`

For the purpose of backwards compatibility, the now removed GC modes can
still be tested for with `compileOption`.

### Removal of `--gc:none`

It is currently not clear how `--gc:none` should work under the new
runtime, as no traditional GC that can be disabled by simply not running
it exists. Sequences and strings are also no longer ref-based, making it
unclear how these should work.

For the time being, the option is removed, but with the intention to
bring it back once the semantics are decided.