ast strictly define nodes with and without sons (#701)
## Summary

Strictly define which AST node kinds have `sons` and which don't, then
update the compiler as required.

## Details

Update the definition of `nkWithSons` and `nkWithoutSons`, use
`nkWithSons` in the `TNode` variant to ensure this is strictly followed.
Update various compiler code as necessary.

The added benefit of this is that a number `case`s are now now
exhaustive and will surface bugs more quickly. This should also make it
easier to introduce more variants down the line.