sem: refactor `newSymGNode` replaces `newSymG` (#702)
## Summary

Internal refactor replacing `sem.newSymG` with `newSymGNode`, no change
in actual behaviour

## Details

This replaces the `newSymG`, which produced a symbol and effected
errors. `newSymGNode` will return an `nkSym` node on succees and
`nkError` if it failed. The `getDefNameSymOrRecover` proc can be used in
conjunction in order to ensure progress can be made in the presence of
errors.

All call sites updated in such a fashion are now more ready to be
converted to a more `nkError` aware style.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>