sigmatch: properly resolve the formal type
`generateTypeInstance` is not intended to be used when it's not certain
that all generic parameters have a concrete type bound -- it's misuse
led to spurious instantiation failure errors.

It first has to be verified that all used generic parameters resolve to
concrete types, which is for now achieved via first calling
`prepareMetatypeForSigmatch` and then checking that the produced type
is not a generic type.